+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 - Help in reactangle terrain mode only shown in create mode
 - Add units to ui components
-- Update core and ui to 6.0
+- Update core and ui to version 6.0